+++ factories/__init__.py
@@ -1,1 +1,2 @@
+from .instance_generator import InstanceGenerator
 from .solver_selections import SolverFactory