+++ CHANGELOG.md
@@ -11,9 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ortools-fast.py` (initially a testing script), which can now solve instance problem files accurately under 5 seconds for each instance file, including the display of metrics, calculations, violations, reasonings if UNSAT.
 - Proper solution events and outcome logging (saving) to user-designated output text file for CLI.
 - Scrollbar for each constraint analysis panel in GUI under statistics tab.
+- Novel instance generator `factories/instance_generator.py` and its dedicated `generate.py` generator script.
+- New constraints `WangLi`, `SUAL`, and `AssignmentDependent` constraints.
+- Activation and Deactivation for new constraints in `main_cli.py` as CLI arguments.
 
 ### Changed
 - Updated the GUI `views.py` and `controllers.py` to fit in to our new solver solution (`ortools-fast.py` but now in a dynamic and versatile way).
+- Updated `README.md`.
 
 ### Improvement
 - Modified metric information in `metadata.py` and how its handled in the `controllers.py`.